requirements.txt
Flask
flask-cors
requests
beautifulsoup4
pymongo
python-dotenv
geopy
